-- change lastused in struct Image from long to int (caused problems with    SDNA on 64-bit platforms)

+++ source/blender/makesdna/DNA_image_types.h
@@ -62,7 +62,7 @@ typedef struct Image {
 	struct PackedFile * packedfile;
 
 	float lastupdate;
-	long  lastused;
+	int lastused;
 	short animspeed;
 	short reserved1;
 	int   reserved2;
